salvando

+++ app/Http/Controllers/MarcaController.php
@@ -7,6 +7,11 @@ use Illuminate\Http\Request;
 
 class MarcaController extends Controller
 {
+    public function __construct(Marca $marca)
+    {
+        $this->marca = $marca;
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -14,9 +19,11 @@ class MarcaController extends Controller
      */
     public function index()
     {
-        $marcas = Marca::all();
+        // $marcas = Marca::all();
+        $marcas = $this->marca->all();
         
-        return $marcas;
+        return response()
+            ->json($marcas, 200);
     }
 
     /**
@@ -27,9 +34,11 @@ class MarcaController extends Controller
      */
     public function store(Request $request)
     {
-        $marca = Marca::create($request->all());
-        
-        return $marca;
+        // $marca = Marca::create($request->all());
+        $marca = $this->marca->create($request->all());
+
+        return response()
+            ->json($marca, 201);
     }
 
     /**
@@ -38,9 +47,21 @@ class MarcaController extends Controller
      * @param  \App\Models\Marca  $marca
      * @return \Illuminate\Http\Response
      */
-    public function show(Marca $marca)
+    public function show($id)
     {
-        return $marca;
+        $marca = $this->marca->find($id);
+
+        if ($marca === null) {
+            // return [
+            //     'erro' => 'Recurso pesquisado não existe'
+            // ];
+
+            return response()
+                ->json(['erro' => 'Recurso pesquisado não existe'], 404);
+        }
+
+        return response()
+            ->json($marca, 200);
     }
 
     /**
@@ -50,11 +71,19 @@ class MarcaController extends Controller
      * @param  \App\Models\Marca  $marca
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, Marca $marca)
+    public function update(Request $request, $id)
     {
+        $marca = $this->marca->find($id);
+
+        if ($marca === null) {
+            return response()
+                ->json(['erro' => 'Impossível realizar a atualização. Recurso pesquisado não existe'], 404);
+        }
+
         $marca->update($request->all());
 
-        return $marca;
+        return response()
+            ->json($marca, 200);
     }
 
     /**
@@ -63,9 +92,18 @@ class MarcaController extends Controller
      * @param  \App\Models\Marca  $marca
      * @return \Illuminate\Http\Response
      */
-    public function destroy(Marca $marca)
+    public function destroy($id)
     {
+        $marca = $this->marca->find($id);
+
+        if ($marca === null) {
+            return response()
+                ->json(['erro' => 'Impossível realizar a exclusão. Recurso pesquisado não existe'], 404);
+        }
+
         $marca->delete();
-        return ['msg' => 'A marca foi removida com sucesso.'];
+
+        return response()
+            ->json(['msg' => 'A marca foi removida com sucesso.'], 200);
     }
 }
